+++ sketch_oxide/src/streaming/mod.rs
@@ -11,12 +11,14 @@
 mod ecm_sketch;
 mod eh_core;
 mod exponential_histogram;
+mod forward_decay;
 mod sliding_hll;
 mod sliding_window;
 mod windowed_aggregator;
 
 pub use ecm_sketch::EcmSketch;
 pub use exponential_histogram::ExponentialHistogram;
+pub use forward_decay::ForwardDecay;
 pub use sliding_hll::{SlidingHLLStats, SlidingHyperLogLog};
 pub use sliding_window::SlidingWindowCounter;
 pub use windowed_aggregator::WindowedAggregator;